Move from checkfunctions to checktype
