+++ lib/checkfunctions.h
@@ -63,7 +63,6 @@ public:
 
         checkFunctions.checkIgnoredReturnValue();
         checkFunctions.checkMissingReturn();  // Missing "return" in exit path
-        checkFunctions.checkIteratorTypeMismatch();
 
         // --check-library : functions with nonmatching configuration
         checkFunctions.checkLibraryMatchFunctions();
@@ -92,8 +91,6 @@ public:
     /** @brief %Check for ignored return values. */
     void checkIgnoredReturnValue();
 
-    void checkIteratorTypeMismatch();
-
     /** @brief %Check for parameters given to math function that do not make sense*/
     void checkMathFunctions();
 
@@ -117,7 +114,6 @@ private:
     void invalidFunctionArgBoolError(const Token *tok, const std::string &functionName, int argnr);
     void invalidFunctionArgStrError(const Token *tok, const std::string &functionName, nonneg int argnr);
     void ignoredReturnValueError(const Token* tok, const std::string& function);
-    void mismatchingIteratorAssignmentType(const Token* tok, const std::string& var_type, const std::string& assigned_type);
     void ignoredReturnErrorCode(const Token* tok, const std::string& function);
     void mathfunctionCallWarning(const Token *tok, const nonneg int numParam = 1);
     void mathfunctionCallWarning(const Token *tok, const std::string& oldexp, const std::string& newexp);
